V1.1 of Quiz App.
Cheats working.
Questions working.
Buttons working.

--------READ ME--------
- Rotation bug needs fixing
This is a weird bug which is related to onCreate in MainActivity. Basically, when a new view is inflated, the question comes up as blank, even though it should show question 1 at index 0. Rotation does retain the question at the correct value, but the question disappears. Test this by going to any question after 1, lets say 3, rotate, go back and next, and it will show question 3, not 1, so it is retaining the value.